+++ x/vesting/types/interfaces.go
@@ -9,6 +9,7 @@ import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	v1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
@@ -74,4 +75,11 @@ type StakingKeeper interface {
 // requires for clawing back unvested coins to the community pool.
 type DistributionKeeper interface {
 	FundCommunityPool(ctx sdk.Context, amount sdk.Coins, sender sdk.AccAddress) error
+}
+
+// GovKeeper defines the expected interface contract the vesting module requires
+// for accessing governance related information.
+type GovKeeper interface {
+	GetParams(ctx sdk.Context) v1.Params
+	GetProposal(ctx sdk.Context, proposalID uint64) (v1.Proposal, bool)
 }